Add multi-arch cross-platform native build support
Add x86/x64 solution and project platform support and integrate native build variants. The WinForms solution and csproj now declare x86/x64 platforms, set PlatformTarget/RuntimeIdentifier, and provide NativeBuildSubdir/NativeCc/NativeAr/NativeMakeArguments plus paths to kernel_loader binaries so native DLLs/libs are built and copied per-arch into the managed target. Makefiles for aeslib, crypto_aes_opencl and kernel_loader were refactored to use BUILD_SUBDIR, per-platform OBJ/BIN/LIB dirs, Windows path conversion, configurable CC/AR/MARCH/OPENCL target, stricter warning options, and improved kernel copying and clean targets. Also small test fixes (explicit unsigned casts) and removal of ARCHITECTURE_NOTE.txt. These changes make building and packaging native dependencies consistent across platforms and architectures.

+++ Feleves_projekt_AES/aeslib/tests/test_aes.c
@@ -452,7 +452,7 @@ static int test_gcm_vectors(void)
             iv2[i] = (uint8_t)(0xa0 + i);
         }
         for (int i = 0; i < 37; i++) {
-            aad2[i] = (uint8_t)(i * 3u + 7u);
+            aad2[i] = (uint8_t)(((unsigned)i * 3u) + 7u);
         }
         for (int i = 0; i < 103; i++) {
             pt2[i] = (uint8_t)(i ^ (i >> 2));
@@ -494,7 +494,7 @@ static int test_gcm_vectors(void)
             iv2[i] = (uint8_t)(0xb0 + i);
         }
         for (int i = 0; i < 37; i++) {
-            aad2[i] = (uint8_t)(i * 5u + 1u);
+            aad2[i] = (uint8_t)(((unsigned)i * 5u) + 1u);
         }
         for (int i = 0; i < 103; i++) {
             pt2[i] = (uint8_t)(i ^ (i >> 1));
